Fixes for svn builds.
The section SB's would fail to create metafiles for packages built from 
the svn sources because the grep "^VERSION=.*" would return a string 
with $SVN_REV in it, which obviously wouldn't match the actual 
(expanded) version string of the package.

As i'd previously added the svn packages to gsb_update, it was easy to 
make gsb_update update the VERSION during it's run, so the metafiles 
should build OK now.

NOTE: Because of this change, when the svn sources are updated in the 
package directories, the version number in the coresponding .pm file for 
that source MUST also be changed - just like you'd have to do with any 
other build.  There is a README file in each package directory that uses 
svn sources to inform people of this.