patch gksu-polkit with upstream git patch.
From d7e31609efe8031dbd56b87caf4200330594488f Mon Sep 17 00:00:00 2001
From: Vincent Untz <vuntz@gnome.org>
Date: Thu, 9 Sep 2010 12:50:38 +0200
Subject: [PATCH] Do not take 100% of the CPU in gksu-polkit when there's no stdin

If there is not stdin (or if it gets closed), we get a EOF, and we
should just stop trying to receive input, instead of trying again and
again.

Also, do not queue anything in the pipe to the stdin of the child if
what we got through our stdin is empty.